Added tests & updated auth.db used by tests
